修改用户表，提现数量统计，加订单目标数量与序号

+++ ruoyi-system/src/main/java/com/ruoyi/business/service/IOrderReceiveRecordService.java
@@ -74,7 +74,7 @@ public interface IOrderReceiveRecordService
      * @param
      * @return 结果
      */
-    long countNumByUserDate();
+    int countNumByUserDate();
 
     /**
      * 前台用户点击后添加订单
